Improve exceptions by giving them methods to get the invalid things
InvalidKeyException now has a constructor taking a key, and the reason
why its invalid.
Also it has now a getKey method to get that key, and getReason to get
the given reason.
InvalidTypeException has now a getExpected method to get the expected
type, and a getReceived method to get the received type.