Only notify Config callback if any config values actually changed
This means it will no longer report changes made by the Config object,
and it will no longer report changes to files that aren't config files.
Make Config#readConfig and Config#readConfigFile return whether any
config value changed
Config now throws exceptions in some cases, rather then just writing a
message to the log. These are all RuntimeExceptions to make it
compatible with old code(and because it would be way to inconvenient
otherwise), but they are documented in the javadoc.