Make Config file writing in createConfig synchronized to the file
This way there should be way less issues when trying to write the same
file from two threads.
Use a copy of the config values list for the file writing, because
writing to files is slow, and that way there should be less
ConcurrentModificationExceptions
Only create config file if it doesn't exist, or is not a file.